Switch to emulated eeprom in flash

+++ src/storage.c
@@ -58,12 +58,12 @@ static const uint8_t storage_default_hoptable[] =
       {0x0,0x1,0x2,0x3,0x4,0x5,0x6,0x7,0x8,0x9,0xA,0xB,0xC,0xD,0xE,0xF,0x10,0x11,0x12,0x13,0x14,0x15,0x16,0x17,0x18,0x19,0x1A,0x1B,0x1C,0x1D,0x1E,0x1F,0x20,0x21,0x22,0x23,0x24,0x25,0x26,0x27,0x28,0x29,0x2A,0x2B,0x2C,0x2D,0x2E,0x2F,0x30,0x31};
 
 void storage_read_from_flash(void) {
-    uint8_t *storage_ptr;
+    uint16_t *storage_ptr;
     uint16_t len;
     uint8_t i;
 
     debug("storage: reading\n"); debug_flush();
-    storage_ptr = (uint8_t*)&storage;
+    storage_ptr = (uint16_t *)&storage;
     len = sizeof(storage);
 
     hal_storage_read(storage_ptr, len);
@@ -94,13 +94,13 @@ void storage_read_from_flash(void) {
 }
 
 void storage_write_to_flash(void) {
-    uint8_t *storage_ptr;
+    uint16_t *storage_ptr;
     uint16_t len;
 
     debug("storage: writing\n"); debug_flush();
     storage.version = STORAGE_VERSION_ID;
 
-    storage_ptr = (uint8_t*)&storage;
+    storage_ptr = (uint16_t*)&storage;
     len = sizeof(storage);
 
     // execute flash write:
